agents/requirements.txt
# Google Agent Development Kit (ADK) Core
google-cloud-aiplatform[adk,agent-engines]>=1.93.0
google-genai>=1.9.0
google-adk>=1.0.0

# Google Cloud Platform Services
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0
google-cloud-logging>=3.5.0

# Data & ML
pydantic>=2.10.6
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.31.0
tenacity>=8.2.3

# Development & Testing
pytest>=8.3.2
pytest-asyncio>=0.23.7
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.4.6
mypy>=1.15.0

# Evaluation
google-adk[eval]>=1.0.0
nest-asyncio>=1.6.0
